refactor in Sound
SoundPlayer::playSpatialSound() does not have to do static_cast on getSpatialVolume() returned value.
All cast things are handled in SpatializationManager.

+++ src/Audio/Sound/Spatialization/spatializationManager.hpp
@@ -10,7 +10,7 @@ namespace Audio {
 class SpatializationManager
 {
 public:
-	int getSpatialVolume(SoundData, const sf::Vector2f& soundPosition, float volume);
+	float getSpatialVolume(SoundData, const sf::Vector2f& soundPosition, float volume);
 	void setListenerPosition(sf::Vector2f listenerPosition) { mListenerPosition = listenerPosition; }
 
 private:
